+++ src/core/object/proxy-clone/CHANGELOG.md
@@ -9,6 +9,13 @@ Changelog
 > - :house:      [Internal]
 > - :nail_care:  [Polish]
 
+## v3.73.0 (2021-12-14)
+
+#### :bug: Bug Fix
+
+* Fixed removing of properties
+* Now the clone function supports descriptors too
+
 ## v3.72.0 (2021-12-10)
 
 #### :rocket: New Feature